python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cmdagent-client"
version = "0.1.0"
description = "Python SDK for cmdagentd"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "grpcio>=1.66.0",
  "protobuf>=4.25.0",
]

[project.optional-dependencies]
ansible = ["ansible-core>=2.17"]
robot = ["robotframework>=7.0"]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["cmdagent_client*"]
